pyproject.toml
[project]
name = "x-poster"
version = "0.1.0"
description = "CLI utility for posting queued text + images to X API v2."
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.24.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "typer>=0.9.0",
]

[project.scripts]
xposter = "xposter.cli:app"
